+++ src/main/java/com/example/demo/models/DTOs/StreetDTO.java
@@ -6,11 +6,11 @@ import jakarta.validation.constraints.NotBlank;
 
 public class StreetDTO {
     @JsonProperty("id")
-    public Integer id;
+    public Long id;
     @JsonProperty("streetName")
     @NotBlank(message = "please specify the street name")
     public String streetName;
     @JsonProperty("areaId")
     @Min(1)
-    public Integer areaId;
+    public Long areaId;
 }
